Fixed the Mutate Function, Added infinite Traffic

+++ utils.py
@@ -3,7 +3,7 @@ import math
 translate_y = 0
 
 def lerp(A, B, t):
-    return int(A +(B-A)*t)
+    return (A +(B-A)*t)
 
 def translateY(y):
     return y + translate_y
